+++ lambda/handler.py
@@ -1,7 +1,12 @@
 """AWS Lambda handler."""
 
+import logging
+
 from mangum import Mangum
 
 from titiler.main import app
 
-handler = Mangum(app, enable_lifespan=False, log_level="error")
+logging.getLogger("mangum.lifespan").setLevel(logging.ERROR)
+logging.getLogger("mangum.http").setLevel(logging.ERROR)
+
+handler = Mangum(app, lifespan="auto", log_level="error")