+++ stack/app.py
@@ -5,6 +5,7 @@ from typing import Any, List, Optional, Union
 
 import docker
 from aws_cdk import aws_apigatewayv2 as apigw
+from aws_cdk import aws_apigatewayv2_integrations as apigw_integrations
 from aws_cdk import aws_ec2 as ec2
 from aws_cdk import aws_ecs as ecs
 from aws_cdk import aws_ecs_patterns as ecs_patterns
@@ -45,6 +46,7 @@ class titilerLambdaStack(core.Stack):
         id: str,
         memory: int = 1024,
         timeout: int = 30,
+        runtime: aws_lambda.Runtime = aws_lambda.Runtime.PYTHON_3_8,
         concurrent: Optional[int] = None,
         permissions: Optional[List[iam.PolicyStatement]] = None,
         layer_arn: Optional[str] = None,
@@ -63,7 +65,7 @@ class titilerLambdaStack(core.Stack):
         lambda_function = aws_lambda.Function(
             self,
             f"{id}-lambda",
-            runtime=aws_lambda.Runtime.PYTHON_3_7,
+            runtime=runtime,
             code=self.create_package(code_dir),
             handler="handler.handler",
             memory_size=memory,
@@ -71,20 +73,6 @@ class titilerLambdaStack(core.Stack):
             timeout=core.Duration.seconds(timeout),
             environment=lambda_env,
         )
-
-        # # If you use dynamodb mosaic backend you should add IAM roles to read/put Item and maybe create Table
-        # permissions.append(
-        #     iam.PolicyStatement(
-        #         actions=[
-        #             "dynamodb:GetItem",
-        #             "dynamodb:PutItem",
-        #             "dynamodb:CreateTable",
-        #             "dynamodb:Scan",
-        #             "dynamodb:BatchWriteItem",
-        #         ],
-        #         resources=[f"arn:aws:dynamodb:{self.region}:{self.account}:table/*"],
-        #     )
-        # )
 
         for perm in permissions:
             lambda_function.add_to_role_policy(perm)
@@ -99,7 +87,9 @@ class titilerLambdaStack(core.Stack):
         api = apigw.HttpApi(
             self,
             f"{id}-endpoint",
-            default_integration=apigw.LambdaProxyIntegration(handler=lambda_function),
+            default_integration=apigw_integrations.LambdaProxyIntegration(
+                handler=lambda_function
+            ),
         )
         core.CfnOutput(self, "Endpoint", value=api.url)
 
@@ -111,7 +101,7 @@ class titilerLambdaStack(core.Stack):
         print("Building docker image...")
         client.images.build(
             path=code_dir,
-            dockerfile="Dockerfiles/lambda/Dockerfile",
+            dockerfile="Dockerfiles/lambda.package",
             tag="titiler-lambda:latest",
             rm=True,
         )
@@ -188,20 +178,6 @@ class titilerECSStack(core.Stack):
             ),
         )
 
-        # # If you use dynamodb mosaic backend you should add IAM roles to read/put Item and maybe create Table
-        # permissions.append(
-        #     iam.PolicyStatement(
-        #         actions=[
-        #             "dynamodb:GetItem",
-        #             "dynamodb:PutItem",
-        #             "dynamodb:CreateTable",
-        #             "dynamodb:Scan",
-        #             "dynamodb:BatchWriteItem",
-        #         ],
-        #         resources=[f"arn:aws:dynamodb:{self.region}:{self.account}:table/*"],
-        #     )
-        # )
-
         for perm in permissions:
             fargate_service.task_definition.task_role.add_to_policy(perm)
 
@@ -239,9 +215,26 @@ if settings.buckets:
     perms.append(
         iam.PolicyStatement(
             actions=["s3:GetObject", "s3:HeadObject"],
-            resources=[f"arn:aws:s3:::{bucket}" for bucket in settings.buckets],
+            resources=[f"arn:aws:s3:::{bucket}*" for bucket in settings.buckets],
         )
     )
+
+# # If you use dynamodb mosaic backend you should add IAM roles to read/put Item and maybe create Table
+# stack = core.Stack()
+# perms.append(
+#     iam.PolicyStatement(
+#         actions=[
+#             "dynamodb:GetItem",
+#             "dynamodb:Scan",
+#             "dynamodb:PutItem",         # Write
+#             "dynamodb:CreateTable",     # Write
+#             "dynamodb:BatchWriteItem",  # Write
+#             "dynamodb:DescribeTable",   # Write
+#         ],
+#         resources=[f"arn:aws:dynamodb:{stack.region}:{stack.account}:table/*"],
+#     )
+# )
+
 
 # Tag infrastructure
 for key, value in {