+++ tests/routes/test_mosaic.py
@@ -98,17 +98,7 @@ def test_info(app):
     assert body["minzoom"] == 7
     assert body["maxzoom"] == 9
     assert body["name"] == "mosaic"  # mosaic.name is not set
-    assert body["quadkeys"] == [
-        "0302300",
-        "0302301",
-        "0302310",
-        "0302302",
-        "0302303",
-        "0302312",
-        "0302320",
-        "0302321",
-        "0302330",
-    ]
+    assert body["quadkeys"] == []
 
 
 def test_tilejson(app):
@@ -187,7 +177,7 @@ def test_tile(app):
             params={"url": MOSAICJSON_FILE},
         )
         assert response.status_code == 200
-        assert response.headers["content-type"] == "image/tiff"
+        assert response.headers["content-type"] == "image/tiff; application=geotiff"
         meta = parse_img(response.content)
         assert meta["width"] == meta["height"] == 256
         assert meta["crs"] == 3857
@@ -197,7 +187,7 @@ def test_tile(app):
             params={"url": MOSAICJSON_FILE, "nodata": 0, "bidx": 1},
         )
         assert response.status_code == 200
-        assert response.headers["content-type"] == "image/tiff"
+        assert response.headers["content-type"] == "image/tiff; application=geotiff"
         meta = parse_img(response.content)
         assert meta["dtype"] == "uint16"
         assert meta["count"] == 2
@@ -229,7 +219,7 @@ def test_tile(app):
             params={"url": MOSAICJSON_FILE, "resampling_method": "bilinear"},
         )
         assert response.status_code == 200
-        assert response.headers["content-type"] == "image/tiff"
+        assert response.headers["content-type"] == "image/tiff; application=geotiff"
 
 
 def test_wmts(app):