+++ titiler/dependencies.py
@@ -3,10 +3,10 @@
 import re
 from dataclasses import dataclass, field
 from enum import Enum
-from typing import Dict, Optional, Tuple, Union
+from typing import Dict, List, Optional, Tuple, Union
 
 import numpy
-from morecantile import tms as DefaultTileMatrixSets
+from morecantile import tms
 from morecantile.models import TileMatrixSet
 from rasterio.enums import Resampling
 from rio_tiler.colormap import cmap
@@ -21,16 +21,9 @@ from starlette.requests import Request
 
 ################################################################################
 #                       CMAP AND TMS Customization
-DefaultTileMatrixSets.register(custom_tms.EPSG3413)
-DefaultTileMatrixSets.register(custom_tms.EPSG6933)
-# REGISTER CUSTOM TMS
-#
-# e.g DefaultTileMatrixSets.register(custom_tms.my_custom_tms)
+tms = tms.register([custom_tms.EPSG3413, custom_tms.EPSG6933])
 
-cmap.register("above", custom_colormap.above_cmap)
-# REGISTER CUSTOM COLORMAP HERE
-#
-# e.g cmap.register("customRed", custom_colormap.custom_red)
+cmap = cmap.register({"above": custom_colormap.above_cmap})
 
 ################################################################################
 # DO NOT UPDATE
@@ -45,7 +38,7 @@ WebMercatorTileMatrixSetName = Enum(  # type: ignore
     "WebMercatorTileMatrixSetName", [("WebMercatorQuad", "WebMercatorQuad")]
 )
 TileMatrixSetNames = Enum(  # type: ignore
-    "TileMatrixSetNames", [(a, a) for a in sorted(DefaultTileMatrixSets.list())]
+    "TileMatrixSetNames", [(a, a) for a in sorted(tms.list())]
 )
 
 
@@ -61,7 +54,7 @@ def WebMercatorTMSParams(
     )
 ) -> TileMatrixSet:
     """TileMatrixSet Dependency."""
-    return DefaultTileMatrixSets.get(TileMatrixSetId.name)
+    return tms.get(TileMatrixSetId.name)
 
 
 def TMSParams(
@@ -71,7 +64,7 @@ def TMSParams(
     )
 ) -> TileMatrixSet:
     """TileMatrixSet Dependency."""
-    return DefaultTileMatrixSets.get(TileMatrixSetId.name)
+    return tms.get(TileMatrixSetId.name)
 
 
 @dataclass
@@ -241,8 +234,13 @@ class RenderParams(DefaultDependency):
         None, description="rio-tiler's colormap name"
     )
     return_mask: bool = Query(True, description="Add mask to the output data.")
+
     colormap: Optional[Dict[int, Tuple[int, int, int, int]]] = field(init=False)
+    rescale_range: Optional[List[Union[float, int]]] = field(init=False)
 
     def __post_init__(self):
         """Post Init."""
         self.colormap = cmap.get(self.color_map.value) if self.color_map else None
+        self.rescale_range = (
+            list(map(float, self.rescale.split(","))) if self.rescale else None
+        )