+++ titiler/settings.py
@@ -9,6 +9,11 @@ class ApiSettings(pydantic.BaseSettings):
     name: str = "titiler"
     cors_origins: str = "*"
     cachecontrol: str = "public, max-age=3600"
+    debug: bool = False
+
+    disable_cog: bool = False
+    disable_stac: bool = False
+    disable_mosaic: bool = False
 
     @pydantic.validator("cors_origins")
     def parse_cors_origin(cls, v):